moved sm510/11/12 specifics to their own files
