+++ src/emu/cpu/sm510/sm510.h
@@ -49,6 +49,7 @@ protected:
 	virtual UINT32 execute_input_lines() const { return 1; }
 	//virtual void execute_set_input(int line, int state);
 	virtual void execute_run();
+	virtual void execute_one() { } // -> child class
 
 	// device_memory_interface overrides
 	virtual const address_space_config *memory_space_config(address_spacenum spacenum = AS_0) const { return(spacenum == AS_PROGRAM) ? &m_program_config : ((spacenum == AS_DATA) ? &m_data_config : NULL); }
@@ -56,7 +57,6 @@ protected:
 	// device_disasm_interface overrides
 	virtual UINT32 disasm_min_opcode_bytes() const { return 1; }
 	virtual UINT32 disasm_max_opcode_bytes() const { return 2; }
-	virtual offs_t disasm_disassemble(char *buffer, offs_t pc, const UINT8 *oprom, const UINT8 *opram, UINT32 options);
 
 	address_space_config m_program_config;
 	address_space_config m_data_config;
@@ -88,7 +88,7 @@ protected:
 
 	// misc internal helpers
 	void increment_pc();
-	virtual void get_opcode_param();
+	virtual void get_opcode_param() { } // -> child class
 
 	UINT8 ram_r();
 	void ram_w(UINT8 data);
@@ -156,14 +156,42 @@ protected:
 	void op_illegal();
 };
 
+
 class sm510_device : public sm510_base_device
 {
 public:
 	sm510_device(const machine_config &mconfig, const char *tag, device_t *owner, UINT32 clock);
+
+protected:
+	virtual offs_t disasm_disassemble(char *buffer, offs_t pc, const UINT8 *oprom, const UINT8 *opram, UINT32 options);
+	virtual void execute_one();
+	virtual void get_opcode_param();
 };
 
 
+class sm511_device : public sm510_base_device
+{
+public:
+	sm511_device(const machine_config &mconfig, const char *tag, device_t *owner, UINT32 clock);
+	sm511_device(const machine_config &mconfig, device_type type, const char *name, const char *tag, device_t *owner, UINT32 clock, int stack_levels, int prgwidth, address_map_constructor program, int datawidth, address_map_constructor data, const char *shortname, const char *source);
+
+protected:
+	virtual offs_t disasm_disassemble(char *buffer, offs_t pc, const UINT8 *oprom, const UINT8 *opram, UINT32 options);
+	virtual void execute_one();
+	virtual void get_opcode_param();
+};
+
+class sm512_device : public sm511_device
+{
+public:
+	sm512_device(const machine_config &mconfig, const char *tag, device_t *owner, UINT32 clock);
+};
+
+
+
 extern const device_type SM510;
+extern const device_type SM511;
+extern const device_type SM512;
 
 
 #endif /* _SM510_H_ */
